+++ src/Farmer/Arm/App.fs
@@ -2,12 +2,14 @@
 module Farmer.Arm.App
 
 open Farmer.ContainerApp
+open Farmer.Identity
 open Farmer
 
 let containerApps = ResourceType ("Microsoft.App/containerApps", "2022-01-01-preview")
 let managedEnvironments = ResourceType ("Microsoft.App/managedEnvironments", "2022-01-01-preview")
 
 open Farmer.ContainerAppValidation
+open Farmer.Identity
 
 type Container =
     { Name : string
@@ -19,6 +21,7 @@ type ContainerApp =
       ActiveRevisionsMode : ActiveRevisionsMode
       IngressMode : IngressMode option
       ScaleRules : Map<string, ScaleRule>
+      Identity: ManagedIdentity
       Replicas : {| Min : int; Max : int |} option
       DaprConfig : {| AppId : string |} option
       Secrets : Map<ContainerAppSettingKey, SecretValue>
@@ -27,6 +30,14 @@ type ContainerApp =
       Containers : Container list
       Location : Location
       Dependencies : Set<ResourceId> }
+    member private this.dependencies = [
+        yield this.Environment
+        yield! this.Dependencies
+        yield! this.Identity.Dependencies
+    ]
+
+    member private this.ResourceId = containerApps.resourceId this.Name
+    member this.SystemIdentity = SystemIdentity this.ResourceId
 
     interface IParameters with
         member this.SecureParameters = [
@@ -44,9 +55,11 @@ type ContainerApp =
     interface IArmResource with
         member this.ResourceId = containerApps.resourceId this.Name
         member this.JsonModel =
-            let dependencies = this.Dependencies.Add this.Environment
-            {| containerApps.Create(this.Name, this.Location, dependencies) with
+            {| containerApps.Create(this.Name, this.Location, this.dependencies) with
                    kind = "containerapp"
+                   identity =
+                       if this.Identity = ManagedIdentity.Empty then Unchecked.defaultof<_>
+                       else this.Identity.ToArmJson
                    properties =
                        {|
                            managedEnvironmentId = this.Environment.Eval()